fix(gpio): change IRQPriority type to uint32_t to prevent signed overflow

+++ drivers/Src/stm32f446xx_gpio_driver.c
@@ -229,7 +229,7 @@ void GPIO_IRQInterruptConfig(uint8_t IRQNumber, uint8_t EnOrDi){
 	}
 }
 
-void GPIO_IRQPriorityConfig(uint8_t IRQNumber, uint8_t IRQPriority){
+void GPIO_IRQPriorityConfig(uint8_t IRQNumber, uint32_t IRQPriority){
 	uint8_t iprx = IRQNumber / 4;
 	uint8_t iprx_section = IRQNumber % 4;
 	uint8_t shift_amount = (8 * iprx_section + (8-NO_PR_BITS_IMPLEMENTED));
